fix: support dual storage strategy.
